protect update and delete actions

+++ src/controllers/ProjectController.ts
@@ -56,6 +56,10 @@ export class ProjectController {
         const error = new Error("Project not found");
         return res.status(404).json({ error: error.message });
       }
+      if (project.manager?.toString() !== req.user?.id.toString()) {
+        const error = new Error('The project can be updated by the manager only')
+        return res.status(404).json({ error: error.message })
+      }
       project.clientName = req.body.clientName
       project.projectName = req.body.projectName
       project.description = req.body.description
@@ -75,8 +79,12 @@ export class ProjectController {
         const error = new Error("Project not found");
         return res.status(404).json({ error: error.message });
       }
+      if (project.manager?.toString() !== req.user?.id.toString()) {
+        const error = new Error('The project can be deleted by the manager only')
+        return res.status(404).json({ error: error.message })
+      }
       await project.deleteOne();
-      res.status(200).json("Project deleted");
+      res.status(200).json("Invalid action");
     } catch (error) {
       console.log(error);
       res.status(500).json({ message: "Error getting project" });
